check syntax and start edit css
