+++ code/src/App.js
@@ -4,8 +4,6 @@ import Allforms from 'components/Allforms';
 
 export const App = () => {
   return (
-    <div>
-      <Allforms />
-    </div>
+    <Allforms />
   );
 }
